refactor piece table and line tracker
- use constrained integer types like u64 for better memory since usize is architecture-dependent.
- update tests and benchmarks according to the changes.
- change split_off for LineTrackerSummary to use split_off() since previous implementation creates iterators needlessly.

+++ crates/mynotes_core/benches/piece_table.rs
@@ -12,7 +12,9 @@ fn bench_insert(c: &mut Criterion) {
             // Insert sequentially.
             // (You could also write a bench for random insertions to force middle-splits!)
             for i in 0..10_000 {
-                table.insert(i, i, 1, BufferKind::Add).unwrap();
+                table
+                    .insert(i as u64, i as u64, 1, BufferKind::Add)
+                    .unwrap(); // UPGRADED to u64
             }
             // black_box prevents the compiler from optimizing away the unused table
             black_box(table);
@@ -24,14 +26,16 @@ fn bench_get(c: &mut Criterion) {
     // 1. Setup a large tree outside the timer
     let mut table = PieceTable::new();
     for i in 0..10_000 {
-        table.insert(i, i, 1, BufferKind::Original).unwrap();
+        table
+            .insert(i as u64, i as u64, 1, BufferKind::Original)
+            .unwrap(); // UPGRADED to u64
     }
 
     c.bench_function("get_piece_middle", |b| {
         b.iter(|| {
             // 2. Measure exactly how long it takes to traverse down to the middle
             // Replace `get` with whatever your actual read/query method is named
-            black_box(table.get_at(5000).unwrap());
+            black_box(table.get_at(5000).unwrap()); // Rust will infer 5000 as u64
         });
     });
 }
@@ -43,14 +47,16 @@ fn bench_delete(c: &mut Criterion) {
                 // Setup: Build a fresh 10k piece tree
                 let mut table = PieceTable::new();
                 for i in 0..10_000 {
-                    table.insert(i, i, 1, BufferKind::Add).unwrap();
+                    table
+                        .insert(i as u64, i as u64, 1, BufferKind::Add)
+                        .unwrap(); // UPGRADED to u64
                 }
                 table
             },
             |mut table| {
                 // Measurement: Delete a massive chunk from the middle.
                 // This will force massive splits, underflows, and merges!
-                table.delete(2500, 5000).unwrap();
+                table.delete(2500, 5000).unwrap(); // Rust will infer u64 here
                 black_box(table);
             },
             // Tells Criterion that the setup phase is expensive and shouldn't be timed
@@ -75,10 +81,13 @@ fn bench_random_insert(c: &mut Criterion) {
                     let insert_pos = if current_length == 0 {
                         0
                     } else {
-                        rng.random_range(0..=current_length)
+                        // UPGRADED range to u64, returning a u64
+                        rng.random_range(0..=current_length as u64)
                     };
 
-                    table.insert(insert_pos, i, 1, BufferKind::Add).unwrap();
+                    table
+                        .insert(insert_pos, i as u64, 1, BufferKind::Add)
+                        .unwrap(); // UPGRADED to u64
                 }
                 black_box(table);
             },
